requirements.txt
Pillow==7.2.0
pytest==6.1.1
